requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
dj-database-url==0.5.0
Django==3.2.13
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
